Merge if, else, elseif nodes into ``if`` node.
I originally had else and elseif designed as separate nodes that would check if they were preceded by an if. The problem with this was that the ``acceptBlock`` function would consume the ending token, leaving no indication of a starting other node like ``elseif`` or ``else. So it just didn't work. One alternative to this would be to just go backward a token afterward, but I opted to just merge the three nodes instead since lua doesn't have something like other languages do with ``else if`` and the three blocks are linked to an ``if`` parent anyways.

Also nicer to have the least amount of nodes to be able to reuse code.