Better benchmarks
LOC is unreliable and the source to parse didn't cover many aspects and harder to parse parts of lua like functions and table literals

Currently tries to do 10kb in 4 seconds on luajit, or 12 seconds on interpreted lua.

So ~2.5kb a second w/ luajit. Pretty decent, would love to optimize this when it covers all of lua properly.